Added text to sections and made goal input inline
